+++ HTMLModule/HTMLPurifier_HTMLModule_Image.m
@@ -23,10 +23,10 @@
         HTMLPurifier_ElementDef* img = [self addElement:@"img" type:@"Inline" contents:@"Empty" attrIncludes:@"Common" attr:@{@"alt*":@"Text", @"height":[NSString stringWithFormat:@"Pixels#%@", max?max:@"1200"], @"width":[NSString stringWithFormat:@"Pixels#%@", max?max:@"1200"], @"longdesc":@"URI", @"src*":[[HTMLPurifier_AttrDef_URI alloc] initWithNumber:@YES]}];
 
         HTMLPurifier_AttrTransform_ImgRequired* transform = [HTMLPurifier_AttrTransform_ImgRequired new];
-        NSString* newKey = [NSString stringWithFormat:@"%ld", img.attr_transform_post.count];
+        NSString* newKey = [NSString stringWithFormat:@"%ld", (unsigned long)img.attr_transform_post.count];
         if (transform && newKey)
             [img.attr_transform_post setObject:transform forKey:newKey];
-        newKey = [NSString stringWithFormat:@"%ld", img.attr_transform_pre.count];
+        newKey = [NSString stringWithFormat:@"%ld", (unsigned long)img.attr_transform_pre.count];
         if (transform && newKey)
             [img.attr_transform_pre setObject:transform forKey:newKey];
     }
